+++ tests/integration/test_attachmentsv2.py
@@ -243,20 +243,14 @@ def test_invalid_item_headers(mini_sentry, relay, invalid_headers, quantity, rea
     assert mini_sentry.get_outcomes(n=2) == [
         {
             "category": DataCategory.ATTACHMENT.value,
-            "org_id": 1,
             "outcome": 3,
-            "key_id": 123,
-            "project_id": 42,
             "reason": reason,
             "quantity": quantity,
             "timestamp": time_within_delta(),
         },
         {
             "category": DataCategory.ATTACHMENT_ITEM.value,
-            "org_id": 1,
             "outcome": 3,
-            "key_id": 123,
-            "project_id": 42,
             "reason": reason,
             "quantity": 1,
             "timestamp": time_within_delta(),
@@ -543,6 +537,7 @@ def test_two_attachments_mapping_to_same_span(mini_sentry, relay):
         }
 
     assert mini_sentry.captured_outcomes.empty()
+    assert mini_sentry.captured_envelopes.empty()
 
 
 @pytest.mark.parametrize(
@@ -605,9 +600,6 @@ def test_span_attachment_ds_drop(mini_sentry, relay, rule_type):
     assert mini_sentry.get_outcomes(n=3) == [
         {
             "timestamp": time_within_delta(),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "Sampled:0",
             "category": DataCategory.ATTACHMENT.value,
@@ -615,9 +607,6 @@ def test_span_attachment_ds_drop(mini_sentry, relay, rule_type):
         },
         {
             "timestamp": time_within_delta(),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "Sampled:0",
             "category": DataCategory.SPAN_INDEXED.value,
@@ -625,9 +614,6 @@ def test_span_attachment_ds_drop(mini_sentry, relay, rule_type):
         },
         {
             "timestamp": time_within_delta(),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "Sampled:0",
             "category": DataCategory.ATTACHMENT_ITEM.value,
@@ -714,9 +700,6 @@ def test_trace_attachment_ds(mini_sentry, relay, rule_type, should_drop):
         assert mini_sentry.get_outcomes(n=2) == [
             {
                 "timestamp": time_within_delta(),
-                "org_id": 1,
-                "project_id": 42,
-                "key_id": 123,
                 "outcome": 1,
                 "reason": "Sampled:0",
                 "category": DataCategory.ATTACHMENT.value,
@@ -724,9 +707,6 @@ def test_trace_attachment_ds(mini_sentry, relay, rule_type, should_drop):
             },
             {
                 "timestamp": time_within_delta(),
-                "org_id": 1,
-                "project_id": 42,
-                "key_id": 123,
                 "outcome": 1,
                 "reason": "Sampled:0",
                 "category": DataCategory.ATTACHMENT_ITEM.value,
@@ -790,9 +770,6 @@ def test_standalone_attachment_only_ds_drop(mini_sentry, relay, rule_type):
     assert mini_sentry.get_outcomes(n=2) == [
         {
             "timestamp": time_within_delta(),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "Sampled:0",
             "category": DataCategory.ATTACHMENT.value,
@@ -800,9 +777,6 @@ def test_standalone_attachment_only_ds_drop(mini_sentry, relay, rule_type):
         },
         {
             "timestamp": time_within_delta(),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "Sampled:0",
             "category": DataCategory.ATTACHMENT_ITEM.value,
@@ -873,9 +847,6 @@ def test_attachments_dropped_with_span_inbound_filters(mini_sentry, relay):
     assert mini_sentry.get_outcomes(n=4) == [
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "release-version",
             "category": DataCategory.ATTACHMENT.value,
@@ -883,9 +854,6 @@ def test_attachments_dropped_with_span_inbound_filters(mini_sentry, relay):
         },
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "release-version",
             "category": DataCategory.SPAN.value,
@@ -893,9 +861,6 @@ def test_attachments_dropped_with_span_inbound_filters(mini_sentry, relay):
         },
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "release-version",
             "category": DataCategory.SPAN_INDEXED.value,
@@ -903,9 +868,6 @@ def test_attachments_dropped_with_span_inbound_filters(mini_sentry, relay):
         },
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 1,
             "reason": "release-version",
             "category": DataCategory.ATTACHMENT_ITEM.value,
@@ -965,9 +927,6 @@ def test_attachment_dropped_with_invalid_spans(mini_sentry, relay):
     assert mini_sentry.get_outcomes(n=4) == [
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 3,
             "reason": "no_data",
             "category": DataCategory.ATTACHMENT.value,
@@ -975,9 +934,6 @@ def test_attachment_dropped_with_invalid_spans(mini_sentry, relay):
         },
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 3,
             "reason": "no_data",
             "category": DataCategory.SPAN.value,
@@ -985,9 +941,6 @@ def test_attachment_dropped_with_invalid_spans(mini_sentry, relay):
         },
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 3,
             "reason": "no_data",
             "category": DataCategory.SPAN_INDEXED.value,
@@ -995,9 +948,6 @@ def test_attachment_dropped_with_invalid_spans(mini_sentry, relay):
         },
         {
             "timestamp": time_within_delta(ts),
-            "org_id": 1,
-            "project_id": 42,
-            "key_id": 123,
             "outcome": 3,
             "reason": "no_data",
             "category": DataCategory.ATTACHMENT_ITEM.value,
@@ -1199,7 +1149,6 @@ def test_span_attachment_independent_rate_limiting(
         outcome_counter[key] += outcome["quantity"]
 
     assert outcome_counter == expected_outcomes
-    assert mini_sentry.captured_outcomes.empty()
 
 
 @pytest.mark.parametrize("owned_by", ["single_span", "multiple_spans", "trace"])